Also accept drives with vFAT label 'cidata' as a configdrive. Match vFAT label using a given label instead of hardcoding 'config-2'.
The NoCloud and ConfigDrive metadata services both attempt to find vFAT
formatted drives but the label used was hardcoded to 'config-2' which
only works with ConfigDrive. BaseConfigDriveService which both inherit
from already has a drive_label property set appropriately for both so
let's use that for the vFAT finding logic as well.